feature/dynamic_resourceDoc: merge DynamicUtilsTest.scala to DynamicUtilTest.scala
