+++ obp-api/src/test/scala/code/util/DynamicUtilTest.scala
@@ -29,22 +29,80 @@ package code.util
 
 import code.api.util._
 import code.setup.PropsReset
+import com.openbankproject.commons.util.{JsonUtils, ReflectUtils}
 import net.liftweb.common.Box
-import org.scalatest.{FeatureSpec, GivenWhenThen, Matchers}
+import net.liftweb.json
+import org.scalatest.{FeatureSpec, FlatSpec, GivenWhenThen, Matchers, Tag}
 
-class DynamicUtilTest extends FeatureSpec with Matchers with GivenWhenThen with PropsReset {
-  
-  feature("test DynamicUtil.compileScalaCode method") {
+class DynamicUtilTest extends FlatSpec with Matchers {
+  object DynamicUtilsTag extends Tag("DynamicUtil")
 
-    
-    scenario("Test the basic function "){
+  implicit val formats = code.api.util.CustomJsonFormats.formats
 
+
+  "DynamicUtil.compileScalaCode method" should "return correct function" taggedAs DynamicUtilsTag in {
       val functionBody: Box[Int => Int] = DynamicUtil.compileScalaCode("def getBank(bankId : Int): Int = bankId+2; getBank _")
       
       val getBankResponse = functionBody.openOrThrowException("")(123)
 
       getBankResponse should be (125)
-    }
   }
 
+
+  val zson = {
+    """
+      |{
+      |   "name": "Sam",
+      |   "age": [12],
+      |   "isMarried": true,
+      |   "weight": 12.11,
+      |   "class": "2",
+      |   "def": 12,
+      |   "email": ["abc@def.com", "hijk@abc.com"],
+      |   "address": [{
+      |     "name": "jieji",
+      |     "code": 123123,
+      |     "street":{"road": "gongbin", "number": 123},
+      |     "_optional_fields_": ["code"]
+      |   }],
+      |   "street": {"name": "hongqi", "width": 12.11},
+      |   "_optional_fields_": ["age", "weight", "address"]
+      |}
+      |""".stripMargin
+  }
+  val zson2 = """{"road": "gongbin", "number": 123}"""
+
+  def buildFunction(jsonStr: String): String => Any = {
+    val caseClasses = JsonUtils.toCaseClasses(json.parse(jsonStr))
+
+    val code =
+      s"""
+         | $caseClasses
+         |
+         | // throws exception: net.liftweb.json.MappingException:
+         | //No usable value for name
+         | //Did not find value which can be converted into java.lang.String
+         |
+         |implicit val formats = code.api.util.CustomJsonFormats.formats
+         |(str: String) => {
+         |  net.liftweb.json.parse(str).extract[RootJsonClass]
+         |}
+         |""".stripMargin
+
+    val fun: Box[String => Any] = DynamicUtil.compileScalaCode(code)
+    fun.orNull
+  }
+
+  "Parse json to dynamic case object" should "success" taggedAs DynamicUtilsTag in {
+
+    val func = buildFunction(zson)
+    val func2 = buildFunction(zson2)
+    val value1 = func.apply(zson)
+    val value2 = func2.apply(zson2)
+
+    ReflectUtils.getNestedField(value1.asInstanceOf[AnyRef], "street", "name") should be ("hongqi")
+    ReflectUtils.getField(value1.asInstanceOf[AnyRef], "weight") shouldEqual Some(12.11)
+    ReflectUtils.getField(value2.asInstanceOf[AnyRef], "number") shouldEqual (123)
+
+  }
 }
